Вносит правки в ДЗ Модуляция
