+++ js/constants.js
@@ -1,0 +1,90 @@
+const PICTURE_COUNT = 25; // ID фото - number
+//const URL_COUNT = 25; // ID url - string
+//const USERS_COUNT = 25; // количество пользователей - number
+const LIKE_MIN_COUNT = 15; // min число лайков - number
+const LIKE_MAX_COUNT = 200; // max число лайков - number
+const COMMENT_COUNT = 500000; // ID комментария - number
+const COMMENTS_MIN_COUNT = 0; // МИН КОЛ-ВО
+const COMMENTS_MAX_COUNT = 20; // МАКС КОЛ-ВО
+const AVATAR_MIN = 1; // min значение
+const AVATAR_MAX = 6; // max значение
+
+// Описания фото
+const DESCRIPTION = [
+  'Начинаем отдыхать!',
+  'Первый закат на пляже',
+  'Наша гостиница',
+  'Вид из окна',
+  'Просто море! Просто отпуск!!!',
+  'И здесь живут люди',
+  'В ресторане',
+  'Красивая девушка',
+  'Просто так, для себя',
+  'Красотища!!!',
+  'Пора домой',
+  'Отпуск закончился',
+  'что то пошло не так ...',
+  'Как вам снимок',
+  'Снова приеду сюда',
+  'Наш номер',
+  'Первый выход на море',
+  'А ещё вот так могу',
+  'Последний кадр',
+  'Фото на память',
+  'Тишина и никаких детей',
+  'Всё серьёзно!',
+  'Не самый хороший мой снимок',
+  'Пробный кадр',
+  'Привет, из дальних стран!',
+  'Хорошо лежать на пляже',
+  'Здравствуй, Родина!',
+  'Вот мы и дома',
+  'Море и небо',
+  'Пора домой',
+];
+
+// Коментарии к фото
+const COMMENT_LINES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+// Имена пользователей, массив
+const USER_NAMES = [
+  'Андрей',
+  'Борис',
+  'Валентин',
+  'Валентина',
+  'Григорий',
+  'Дмитрий',
+  'Дарья',
+  'Егор',
+  'Елена',
+  'Зинаида',
+  'Игорь',
+  'Карина',
+  'Леонид',
+  'Мария',
+  'Николай',
+  'Ольга',
+];
+
+export {
+  PICTURE_COUNT,
+  LIKE_MIN_COUNT,
+  LIKE_MAX_COUNT,
+  COMMENT_COUNT,
+  COMMENTS_MIN_COUNT,
+  COMMENTS_MAX_COUNT,
+  AVATAR_MIN,
+  AVATAR_MAX,
+  DESCRIPTION,
+  COMMENT_LINES,
+  USER_NAMES
+};
+
+
